switches uints to ints

+++ network/p2p/inspector/validation/utils_test.go
@@ -9,14 +9,14 @@ import (
 // TestDuplicateStringTracker tests the duplicateStrTracker.track function.
 func TestDuplicateStringTracker(t *testing.T) {
 	tracker := make(duplicateStrTracker)
-	require.Equal(t, uint(1), tracker.track("test1"))
-	require.Equal(t, uint(2), tracker.track("test1"))
+	require.Equal(t, 1, tracker.track("test1"))
+	require.Equal(t, 2, tracker.track("test1"))
 
 	// tracking a new string, 3 times
-	require.Equal(t, uint(1), tracker.track("test2"))
-	require.Equal(t, uint(2), tracker.track("test2"))
-	require.Equal(t, uint(3), tracker.track("test2"))
+	require.Equal(t, 1, tracker.track("test2"))
+	require.Equal(t, 2, tracker.track("test2"))
+	require.Equal(t, 3, tracker.track("test2"))
 
 	// tracking an empty string
-	require.Equal(t, uint(1), tracker.track(""))
+	require.Equal(t, 1, tracker.track(""))
 }
